+++ BoredGame/src/main/java/com/boredgame/repos/UsersRepos.java
@@ -3,6 +3,6 @@ package com.boredgame.repos;
 import com.boredgame.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepos extends JpaRepository<User, Long> {
+public interface UsersRepos extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 }
